requirements.txt
absl-py==0.8.0
aniso8601==7.0.0
appnope==0.1.0
asn1crypto==0.24.0
astor==0.7.1
astroid==2.2.5
attrs==19.1.0
autopep8==1.4.3
awscli==1.16.232
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
boto==2.49.0
boto3==1.9.225
botocore==1.12.225
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.3.9
cryptography==2.7
decorator==4.4.0
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
firebase-admin==3.0.0
flake8==3.7.7
Flask==1.0.2
flask-restplus==0.13.0
Flask-WTF==0.14.2
gast==0.2.2
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.19.1
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.22.0
gunicorn==19.9.0
h5py==2.9.0
httplib2==0.13.1
idna==2.8
imgurpython==1.1.7
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.16
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10
jmespath==0.9.4
jsonschema==3.0.2
jupyter-client==5.3.1
jupyter-core==4.5.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.3.1
Mako==1.1.0
Markdown==3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mdv==1.7.4
mistune==0.8.4
msgpack==0.6.1
nb-conda==2.2.1
nb-conda-kernels==2.2.2
nbconvert==5.5.0
nbformat==4.4.0
notebook==6.0.0
numpy==1.16.2
olefile==0.46
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.3.1
pygpu==0.7.6
pylint==2.3.1
pymongo==3.8.0
pyOpenSSL==19.0.0
pyrsistent==0.15.4
PySocks==1.7.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-http-client==3.2.1
pytz==2018.9
PyYAML==5.1.2
pyzmq==18.1.0
requests==2.21.0
rsa==3.4.2
s3transfer==0.2.1
scipy==1.3.1
Send2Trash==1.5.0
sendgrid==6.1.0
six==1.12.0
soupsieve==1.9
tabulate==0.8.3
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
Theano==1.0.4
tornado==6.0.3
traitlets==4.3.2
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
wrapt==1.11.1
WTForms==2.2.1
